requirements.txt
pyserial
crcmod
opencv-contrib-python
python-osc